pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-mini-claw"
version = "0.1.0"
description = "Autonomous AI research pipeline (LangGraph + multi-agent + Supabase)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Ungureanu Sergiu-Stefan" }]
license = { text = "MIT" }

dependencies = [
    # Core LLM + orchestration
    "anthropic>=0.39.0",
    "langgraph>=0.2.40",

    # Literature retrieval
    "arxiv>=2.1.3",

    # Sandbox execution
    "docker>=7.1.0",

    # Datasets / models
    "datasets>=3.0.0",
    "huggingface_hub>=0.26.0",

    # Embeddings + ML utilities
    "sentence-transformers>=3.0.0",
    "scikit-learn>=1.5.0",
    "numpy>=1.26.0",

    # CLI / HITL formatting
    "rich>=13.9.0",

    # Supabase persistence
    "supabase>=2.9.0",

    # Configuration
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]

[project.scripts]
mini-claw = "backend.main:cli"

[tool.setuptools.packages.find]
include = ["backend*"]
exclude = ["tests*", "src*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
